METAL-3782 better indication why database creation failed

+++ pkg/utils/database/postgres.go
@@ -41,6 +41,7 @@ type Postgres struct {
 	Database         string
 	User             string
 	Password         string
+	Monitoring       bool
 	Extensions       []string
 	SSLEnabled       bool
 	SkipCAVerify     bool
@@ -161,10 +162,21 @@ func (p Postgres) createDatabase(admin AdminCredentials) error {
 		}
 	}
 
+	if p.Monitoring {
+		err := p.enableMonitoring(admin)
+		if err != nil {
+			return fmt.Errorf("can not enable monitoring - %s", err)
+		}
+	}
+
+	err := p.addExtensions(admin)
+	if err != nil {
+		return fmt.Errorf("can not add extension - %s", err)
+	}
+
 	if p.DropPublicSchema {
 		if err := p.dropPublicSchema(admin); err != nil {
-			logrus.Errorf("failed dropping the public schema %s", err)
-			return err
+			return fmt.Errorf("can not drop public schema - %s", err)
 		}
 		if len(p.Schemas) == 0 {
 			logrus.Info("the public schema is dropped, but no additional schemas are created, schema creation must be handled on the application side now")
@@ -176,12 +188,6 @@ func (p Postgres) createDatabase(admin AdminCredentials) error {
 			logrus.Errorf("failed creating additional schemas %s", err)
 			return err
 		}
-	}
-
-	err := p.addExtensions(admin)
-	if err != nil {
-		logrus.Errorf("failed creating postgres extensions %s", err)
-		return err
 	}
 
 	return nil
@@ -223,6 +229,10 @@ func (p Postgres) createUser(admin AdminCredentials) error {
 }
 
 func (p Postgres) dropPublicSchema(admin AdminCredentials) error {
+	if p.Monitoring {
+		return fmt.Errorf("can not drop public schema when monitoring is enabled on instance level")
+	}
+
 	drop := "DROP SCHEMA IF EXISTS public;"
 	if err := p.executeExec(p.Database, drop, admin); err != nil {
 		logrus.Errorf("failed to drop the schema Public: %s", err)
@@ -321,10 +331,21 @@ func (p Postgres) addExtensions(admin AdminCredentials) error {
 		query := fmt.Sprintf("CREATE EXTENSION IF NOT EXISTS \"%s\";", ext)
 		err := p.executeExec(p.Database, query, admin)
 		if err != nil {
-			logrus.Errorf("failed creating extensions %s - %s", query, err)
 			return err
 		}
 	}
+	return nil
+}
+
+func (p Postgres) enableMonitoring(admin AdminCredentials) error {
+	monitoringExtension := "pg_stat_statements"
+
+	query := fmt.Sprintf("CREATE EXTENSION IF NOT EXISTS \"%s\";", monitoringExtension)
+	err := p.executeExec(p.Database, query, admin)
+	if err != nil {
+		return err
+	}
+
 	return nil
 }
